requirements.txt
./Adafruit_Python_DHT
paho-mqtt
RPi.GPIO
#sysv_ipc
